refactor: split cli into package
