+++ cmd/goxa/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "goxa"
+
+func main() {
+	goxa.Main()
+}
